pyproject.toml
# pyproject.toml
# Docling Hybrid OCR - PDF to Markdown with VLM backends

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docling-hybrid-ocr"
version = "0.1.0"
description = "Hybrid PDF to Markdown conversion using Docling and VLM backends"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Docling Hybrid Team"},
]
keywords = ["pdf", "ocr", "vlm", "markdown", "docling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
]

dependencies = [
    # Core dependencies
    "pydantic>=2.0.0",
    "tomli>=2.0.0; python_version<'3.11'",
    
    # PDF processing
    "docling>=2.0.0",
    "pypdfium2>=4.0.0",
    
    # Async HTTP
    "aiohttp>=3.9.0",
    
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # Logging
    "structlog>=23.0.0",
    
    # Image processing
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-httpx>=0.30.0",
    "aioresponses>=0.7.0",
    
    # Linting & Formatting
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    
    # Type stubs
    "types-Pillow",
]

vllm = [
    # For DeepSeek-OCR via vLLM (CUDA Linux)
    "openai>=1.0.0",  # OpenAI-compatible client
]

mlx = [
    # For DeepSeek-OCR via MLX (macOS)
    "mlx>=0.10.0; sys_platform == 'darwin'",
    "mlx-vlm>=0.1.0; sys_platform == 'darwin'",
]

[project.scripts]
docling-hybrid-ocr = "docling_hybrid.cli.main:app"

[project.urls]
Homepage = "https://github.com/docling-hybrid/docling-hybrid-ocr"
Documentation = "https://github.com/docling-hybrid/docling-hybrid-ocr#readme"
Repository = "https://github.com/docling-hybrid/docling-hybrid-ocr"
Issues = "https://github.com/docling-hybrid/docling-hybrid-ocr/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docling_hybrid"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require external services)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/docling_hybrid"]
omit = ["tests/*", "**/stubs/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["docling_hybrid"]
